// wada-ama-testcenter-web-57718053929f/wada-ama-testcenter-web-57718053929f/src/app/tdp/models/period.enum.ts
export enum PeriodType {
    Yearly = 0,
    Quarterly = 1,
    Monthly = 2,
}

export enum Year {
    Last = 0,
    Current = 1,
    Next = 2,
}

export enum Quarter {
    None = -1,
    Q1 = 0,
    Q2 = 1,
    Q3 = 2,
    Q4 = 3,
}

export enum Month {
    None = -1,
    Jan = 0,
    Feb = 1,
    Mar = 2,
    Apr = 3,
    May = 4,
    Jun = 5,
    Jul = 6,
    Aug = 7,
    Sep = 8,
    Oct = 9,
    Nov = 10,
    Dec = 11,
}